[client] Iterate on snapshot of keys in OpenCTIStix2Splitter.enlist_element
The (key, value) pair iterated across in the for loop in enlist_element
can get out of sync with the data structure, as assignments of the form
dict[key] = value, which are used inside the loop, are not guaranteed to
preserve the [key] instance in-place, which could change the
dictionary's .items() iteration source. To address this, put the keys in
a list(), to take an independent copy of the keys that won't be altered
by edits to the data structure inside of the loop. It does not appear
that any attempts to delete a key occur, just set a value to None, so
taking a copy to iterate off of should be safe.